refactor(mcp): split the transport into reviewable seams
server.ts was 939 lines holding route wiring, authorization, registration,
dispatch, the action-policy gate, and Zod schema projection in one file, behind
an explicit file-size exception. It is the single file that the observability,
progressive-disclosure, response-budget, error, guide-layer, and protocol
workstreams all need to edit, so concurrent work on it would serialize into
merge conflicts.

Extract authorization.ts, register.ts, dispatch.ts, schema-projection.ts,
graph-composition.ts, types.ts, and guards.ts. server.ts keeps route wiring and
graph composition and re-exports the option types, so the package's public
surface is unchanged.

No behaviour change; the existing 34 tests pass untouched apart from one import
path. Largest module is now 279 lines, so the file-size exception is removed.

Refs #3924
