+++ packages/mcp/src/guards.ts
@@ -1,0 +1,4 @@
+/** Narrow type guards shared by the MCP transport modules. */
+export function isRecord(value: unknown): value is Record<string, unknown> {
+  return typeof value === "object" && value !== null && !Array.isArray(value)
+}
